demo: merge client-side tools into AiAgentView (VSDK-253)
Replace the separate AiAgentTabView/AiAgentClientToolsView sub-tab
structure with a single unified AiAgentView that reuses the existing
version picker and all option fields (agentId, version, versionId,
region, conversationId, toggles, call options, custom attributes).

The client-side tools demo (send_message SMS tool) is now a toggle
within the same form:
- When enabled: the widget is embedded directly in the React tree
  (not in an iframe) so client tools can be registered imperatively
- SMS config fields (API key, from number) appear in the form
- Registered tools + tool invocation log cards render below the event log
- All other form fields (version picker, toggles, call options, custom
  attributes) are shared between both embed modes

Removes AiAgentTabView.tsx and AiAgentClientToolsView.tsx.
Reverts App.tsx to import AiAgentView directly.
